moving remap public and adding to meshing
