+++ pychunkedgraph/app/meshing/legacy/routes.py
@@ -3,8 +3,11 @@ from middle_auth_client import auth_requires_permission, auth_required
 
 from pychunkedgraph.app.meshing import common
 from pychunkedgraph.graph import exceptions as cg_exceptions
+from pychunkedgraph.app.app_utils import remap_public
 
-bp = Blueprint("pcg_meshing_v0", __name__, url_prefix=f"/{common.__meshing_url_prefix__}/1.0")
+bp = Blueprint(
+    "pcg_meshing_v0", __name__, url_prefix=f"/{common.__meshing_url_prefix__}/1.0"
+)
 
 # -------------------------------
 # ------ Access control and index
@@ -54,6 +57,7 @@ def api_exception(e):
 
 
 @bp.route("/<table_id>/<node_id>/validfragments", methods=["POST", "GET"])
+@remap_public
 @auth_requires_permission("view")
 def handle_valid_frags(table_id, node_id):
     return common.handle_valid_frags(table_id, node_id)
@@ -64,5 +68,6 @@ def handle_valid_frags(table_id, node_id):
 
 @bp.route("/<table_id>/manifest/<node_id>:0", methods=["GET"])
 @auth_requires_permission("view")
+@remap_public
 def handle_get_manifest(table_id, node_id):
     return common.handle_get_manifest(table_id, node_id)
