+++ pychunkedgraph/app/meshing/v1/routes.py
@@ -3,8 +3,11 @@ from middle_auth_client import auth_requires_permission, auth_required
 
 from pychunkedgraph.app.meshing import common
 from pychunkedgraph.graph import exceptions as cg_exceptions
+from pychunkedgraph.app.app_utils import remap_public
 
-bp = Blueprint("pcg_meshing_v1", __name__, url_prefix=f"/{common.__meshing_url_prefix__}/api/v1")
+bp = Blueprint(
+    "pcg_meshing_v1", __name__, url_prefix=f"/{common.__meshing_url_prefix__}/api/v1"
+)
 
 # -------------------------------
 # ------ Access control and index
@@ -56,6 +59,7 @@ def api_exception(e):
 
 @bp.route("/table/<table_id>/node/<node_id>/validfragments", methods=["GET"])
 @auth_requires_permission("view")
+@remap_public
 def handle_valid_frags(table_id, node_id):
     return common.handle_valid_frags(table_id, node_id)
 
@@ -65,12 +69,16 @@ def handle_valid_frags(table_id, node_id):
 
 @bp.route("/table/<table_id>/manifest/<node_id>:0", methods=["GET"])
 @auth_requires_permission("view")
+@remap_public
 def handle_get_manifest(table_id, node_id):
     return common.handle_get_manifest(table_id, node_id)
 
+
 ## ENQUE MESHING JOBS ----------------------------------------------------------
+
 
 @bp.route("/table/<table_id>/remeshing", methods=["POST"])
 @auth_requires_permission("edit")
+@remap_public(edit=True)
 def handle_remesh(table_id):
     return common.handle_remesh(table_id)
