+++ pychunkedgraph/app/segmentation/common.py
@@ -31,73 +31,10 @@ from pychunkedgraph.graph.attributes import OperationLogs
 from pychunkedgraph.meshing import mesh_analysis
 from pychunkedgraph.graph.misc import get_contact_sites
 from middle_auth_client import get_usernames
-from functools import wraps
-from werkzeug.datastructures import ImmutableMultiDict
+
 
 __api_versions__ = [0, 1]
 __segmentation_url_prefix__ = os.environ.get("SEGMENTATION_URL_PREFIX", "segmentation")
-
-
-def remap_public(func=None, *, edit=False, check_node_ids=False):
-    def mydecorator(f):
-        @wraps(f)
-        def decorated_function(*args, **kwargs):
-            virtual_tables = current_app.config.get("VIRTUAL_TABLES", {})
-            table_id = kwargs.get("table_id")
-            http_args = request.args.to_dict()
-
-            if table_id is None:
-                # then no table remapping necessary
-                return f(*args, **kwargs)
-            if not table_id in virtual_tables:
-                # if table table_id isn't in virtual
-                # tables then just return
-                return f(*args, **kwargs)
-            else:
-                # then we have a virtual table
-                if edit:
-                    raise Exception("No edits allowed on virtual tables")
-                # then we want to remap the table name
-                new_table = virtual_tables[table_id]["table_id"]
-                kwargs["table_id"] = new_table
-                v_timestamp = virtual_tables[table_id]["timestamp"]
-                v_timetamp_float = time.mktime(v_timestamp.timetuple())
-
-                # we want to fix timestamp parameters too
-                http_args["timestamp"] = v_timetamp_float
-                http_args["timestamp_future"] = v_timetamp_float
-                request.args = ImmutableMultiDict(http_args)
-
-                cg = app_utils.get_cg(new_table)
-
-                def assert_node_prop(prop):
-                    node_id = kwargs.get(prop, None)
-                    if node_id is not None:
-                        node_id = int(node_id)
-                        # check if this root_id is valid at this timestamp
-                        timestamp = cg.get_node_timestamps([node_id])
-                        if not np.all(timestamp < np.datetime64(v_timestamp)):
-                            raise Exception("root_id not valid at timestamp")
-
-                assert_node_prop("root_id")
-                assert_node_prop("node_id")
-
-                if check_node_ids:
-                    node_ids = np.array(
-                        json.loads(request.data)["node_ids"], dtype=np.uint64
-                    )
-                    timestamps = cg.get_node_timestamps(node_ids)
-                    if not np.all(timestamps < np.datetime64(v_timestamp)):
-                        raise Exception("node_ids are all not valid at timestamp")
-
-                return f(*args, **kwargs)
-
-        return decorated_function
-
-    if func:
-        return mydecorator(func)
-    else:
-        return mydecorator
 
 
 def index():
