+++ pychunkedgraph/app/segmentation/generic/routes.py
@@ -5,6 +5,7 @@ from middle_auth_client import (
     auth_requires_permission,
 )
 from pychunkedgraph.app.segmentation import common
+from pychunkedgraph.app.app_utils import remap_public
 
 bp = Blueprint(
     "pcg_generic_v1", __name__, url_prefix=f"/{common.__segmentation_url_prefix__}"
@@ -42,7 +43,7 @@ def sleep_me(sleep):
 
 @bp.route("/table/<table_id>/info", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public
+@remap_public
 def handle_info(table_id):
     return common.handle_info(table_id)
