+++ pychunkedgraph/app/segmentation/v1/routes.py
@@ -9,7 +9,7 @@ from middle_auth_client import auth_requires_permission
 from middle_auth_client import auth_requires_admin
 from middle_auth_client import auth_required
 
-from pychunkedgraph.app.app_utils import jsonify_with_kwargs, toboolean, tobinary
+from pychunkedgraph.app.app_utils import jsonify_with_kwargs, toboolean, tobinary, remap_public
 from pychunkedgraph.app.segmentation import common
 from pychunkedgraph.graph import exceptions as cg_exceptions
 
@@ -69,7 +69,7 @@ def api_exception(e):
 
 @bp.route("/table/<table_id>/merge", methods=["POST"])
 @auth_requires_permission("edit")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_merge(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     merge_result = common.handle_merge(table_id)
@@ -82,7 +82,7 @@ def handle_merge(table_id):
 
 @bp.route("/table/<table_id>/merge_admin", methods=["POST"])
 @auth_requires_permission("admin")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_merge_admin(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     allow_same_segment_merge = request.args.get(
@@ -103,7 +103,7 @@ def handle_merge_admin(table_id):
 
 @bp.route("/table/<table_id>/split", methods=["POST"])
 @auth_requires_permission("edit")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_split(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     split_result = common.handle_split(table_id)
@@ -116,7 +116,7 @@ def handle_split(table_id):
 
 @bp.route("/table/<table_id>/graph/split_preview", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_split_preview(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     split_preview = common.handle_split_preview(table_id)
@@ -128,7 +128,7 @@ def handle_split_preview(table_id):
 
 @bp.route("/table/<table_id>/undo", methods=["POST"])
 @auth_requires_permission("edit")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_undo(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     undo_result = common.handle_undo(table_id)
@@ -144,7 +144,7 @@ def handle_undo(table_id):
 
 @bp.route("/table/<table_id>/redo", methods=["POST"])
 @auth_requires_permission("edit")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_redo(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     redo_result = common.handle_redo(table_id)
@@ -160,7 +160,7 @@ def handle_redo(table_id):
 
 @bp.route("/table/<table_id>/rollback_user", methods=["POST"])
 @auth_requires_admin
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_rollback(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     rollback_result = common.handle_rollback(table_id)
@@ -173,7 +173,7 @@ def handle_rollback(table_id):
 
 @bp.route("/table/<table_id>/user_operations", methods=["GET"])
 @auth_requires_permission("admin_view")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def handle_user_operations(table_id):
     disp = request.args.get("disp", default=False, type=toboolean)
     user_operations = pd.DataFrame.from_dict(common.all_user_operations(table_id))
@@ -189,7 +189,7 @@ def handle_user_operations(table_id):
 
 @bp.route("/table/<table_id>/node/<node_id>/root", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public
+@remap_public
 def handle_root(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     root_id = common.handle_root(table_id, node_id)
@@ -202,7 +202,7 @@ def handle_root(table_id, node_id):
 
 @bp.route("/table/<table_id>/roots", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_roots(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     root_ids = common.handle_roots(table_id, is_binary=False)
@@ -220,7 +220,7 @@ def handle_roots(table_id):
 
 @bp.route("/table/<table_id>/roots_binary", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_roots_binary(table_id):
     root_ids = common.handle_roots(table_id, is_binary=True)
     return tobinary(root_ids)
@@ -231,7 +231,7 @@ def handle_roots_binary(table_id):
 
 @bp.route("/table/<table_id>/node/<node_id>/children", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_children(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     children_ids = common.handle_children(table_id, node_id)
@@ -244,7 +244,7 @@ def handle_children(table_id, node_id):
 
 @bp.route("/table/<table_id>/l2_chunk_children/<chunk_id>", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_l2_chunk_children(table_id, chunk_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     as_array = request.args.get("as_array", default=False, type=toboolean)
@@ -261,7 +261,7 @@ def handle_l2_chunk_children(table_id, chunk_id):
 
 @bp.route("/table/<table_id>/l2_chunk_children_binary/<chunk_id>", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_l2_chunk_children_binary(table_id, chunk_id):
     as_array = request.args.get("as_array", default=False, type=toboolean)
     l2_chunk_children = common.handle_l2_chunk_children(table_id, chunk_id, as_array)
@@ -276,7 +276,7 @@ def handle_l2_chunk_children_binary(table_id, chunk_id):
 
 @bp.route("/table/<table_id>/node/<node_id>/leaves", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_leaves(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     leaf_ids = common.handle_leaves(table_id, node_id)
@@ -290,7 +290,7 @@ def handle_leaves(table_id, node_id):
 @bp.route("/table/<table_id>/node/leaves_many", methods=["POST"])
 @bp.route("/table/<table_id>/leaves_many", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(check_node_ids=True)
+@remap_public(check_node_ids=True)
 def handle_leaves_many(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     root_to_leaf_dict = common.handle_leaves_many(table_id)
@@ -302,7 +302,7 @@ def handle_leaves_many(table_id):
 
 @bp.route("/table/<table_id>/node/<node_id>/subgraph", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_subgraph(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     edges = common.handle_subgraph(table_id, node_id)
@@ -319,7 +319,7 @@ def handle_subgraph(table_id, node_id):
 
 @bp.route("/table/<table_id>/node/<node_id>/contact_sites", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_contact_sites(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     contact_sites, contact_site_metadata = common.handle_contact_sites(
@@ -337,7 +337,7 @@ def handle_contact_sites(table_id, node_id):
     methods=["GET"],
 )
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_pairwise_contact_sites(table_id, first_node_id, second_node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     contact_sites, contact_site_metadata = common.handle_pairwise_contact_sites(
@@ -355,7 +355,7 @@ def handle_pairwise_contact_sites(table_id, first_node_id, second_node_id):
 
 @bp.route("/table/<table_id>/change_log", methods=["GET"])
 @auth_requires_admin
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def change_log_full(table_id):
     si = io.StringIO()
     cw = csv.writer(si)
@@ -370,7 +370,7 @@ def change_log_full(table_id):
 
 @bp.route("/table/<table_id>/tabular_change_log_recent", methods=["GET"])
 @auth_requires_permission("admin_view")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def tabular_change_log_weekly(table_id):
     disp = request.args.get("disp", default=False, type=toboolean)
     weekly_tab_change_log = common.tabular_change_log_recent(table_id)
@@ -383,7 +383,7 @@ def tabular_change_log_weekly(table_id):
 
 @bp.route("/table/<table_id>/root/<root_id>/change_log", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=True)
+@remap_public(edit=True)
 def change_log(table_id, root_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     log = common.change_log(table_id, root_id)
@@ -392,7 +392,7 @@ def change_log(table_id, root_id):
 
 @bp.route("/table/<table_id>/root/<root_id>/tabular_change_log", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def tabular_change_log(table_id, root_id):
     disp = request.args.get("disp", default=False, type=toboolean)
     get_root_ids = request.args.get("root_ids", default=False, type=toboolean)
@@ -409,7 +409,7 @@ def tabular_change_log(table_id, root_id):
 
 @bp.route("/table/<table_id>/root/<root_id>/merge_log", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def merge_log(table_id, root_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     log = common.merge_log(table_id, root_id)
@@ -418,7 +418,7 @@ def merge_log(table_id, root_id):
 
 @bp.route("/table/<table_id>/root/<root_id>/lineage_graph", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_lineage_graph(table_id, root_id):
     from networkx import node_link_data
 
@@ -429,7 +429,7 @@ def handle_lineage_graph(table_id, root_id):
 
 @bp.route("/table/<table_id>/past_id_mapping", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_past_id_mapping(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     resp = common.handle_past_id_mapping(table_id)
@@ -438,7 +438,7 @@ def handle_past_id_mapping(table_id):
 
 @bp.route("/table/<table_id>/oldest_timestamp", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def oldest_timestamp(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     delimiter = request.args.get("delimiter", default=" ", type=str)
@@ -449,7 +449,7 @@ def oldest_timestamp(table_id):
 
 @bp.route("/table/<table_id>/root/<root_id>/last_edit", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def last_edit(table_id, root_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     delimiter = request.args.get("delimiter", default=" ", type=str)
@@ -463,7 +463,7 @@ def last_edit(table_id, root_id):
 
 @bp.route("/table/<table_id>/graph/find_path", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def find_path(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     precision_mode = request.args.get("precision_mode", default=True, type=toboolean)
@@ -477,7 +477,7 @@ def find_path(table_id):
 
 @bp.route("/table/<table_id>/is_latest_roots", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_is_latest_roots(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     is_latest_roots = common.handle_is_latest_roots(table_id, is_binary=False)
@@ -502,7 +502,7 @@ def handle_root_timestamps(table_id):
 
 @bp.route("/table/<table_id>/roots_from_coords", methods=["POST"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_roots_from_coords(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     resp = common.handle_roots_from_coord(table_id)
@@ -512,7 +512,7 @@ def handle_roots_from_coords(table_id):
 ## Get level2 graph -------------------------------------------------------------
 @bp.route("/table/<table_id>/node/<node_id>/lvl2_graph", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def handle_get_lvl2_graph(table_id, node_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     resp = common.handle_get_layer2_graph(table_id, node_id)
@@ -524,7 +524,7 @@ def handle_get_lvl2_graph(table_id, node_id):
 
 @bp.route("/table/<table_id>/operation_details", methods=["GET"])
 @auth_requires_permission("view")
-@common.remap_public(edit=False)
+@remap_public(edit=False)
 def operation_details(table_id):
     int64_as_str = request.args.get("int64_as_str", default=False, type=toboolean)
     resp = common.operation_details(table_id)
